requirements.txt
cx-Oracle==7.2.3
Django==1.11.23
django-firebird==1.11
django-redis==4.10.0
django-settings-export==1.2.1
django-tables2==1.13.0
fdb==2.0.1
future==0.18.2
gunicorn==19.9.0
Pillow==6.2.1
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-stdnum==1.13
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
